requirements/base_req.txt
Pillow==9.0.1  # https://github.com/python-pillow/Pillow

# Django
# ------------------------------------------------------------------------------
django==3.2.4  # pyup: < 3.2  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-cors-headers==3.13.0

# Database
# ------------------------------------------------------------------------------
psycopg2-binary

# Deploy
# ------------------------------------------------------------------------------
gunicorn==20.1.0  # https://github.com/benoitc/gunicorn

# Telegram
# ------------------------------------------------------------------------------
aiogram

